Fix `split_string` for unicode strings
